requirements.txt
mysql-connector-python==8.0.32
SQLAlchemy==2.0.27